requirements.txt
aiofiles==23.2.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
bleach==6.1.0
blinker==1.7.0
bokeh==3.3.4
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
fonttools==4.47.2
fsspec==2023.12.2
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gradio==4.16.0
gradio_client==0.8.1
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.21.1
matplotlib==3.8.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
numpy==1.26.3
openai==1.10.0
orjson==3.9.12
packaging==23.2
pandas==2.2.0
panel==1.3.8
param==2.0.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.4
pyviz_comms==3.0.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruff==0.1.15
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.28
starlette==0.35.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
urllib3==2.2.0
uvicorn==0.27.0.post1
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
xyzservices==2023.10.1
